Add support for using embedding models from Databricks. (#45)
Databricks hosts a couple of dedicated text embedding models through
their Foundation Models API [0]. This commit introduces
`embed_databricks()` to give ragnar users access to them.

Note that this is a pretty quick-and-dirty pass. I'm being cheeky here
and hooking into `ellmer` internals to handle all of the credential
stuff, which might not be the cleanest approach in the future.

[0]: https://docs.databricks.com/aws/en/machine-learning/foundation-model-apis/supported-models

Signed-off-by: Aaron Jacobs <aaron.jacobs@posit.co>